9-19-24 [SWGDL] Added petDamageMultiplier / Fixed petOutOfCombatHAMRegenMultiplier
Changed when petAllMountsUsedByAnyone and you are a CH, you can use the pet as normal.   If you are not a CH and you have pet/mount <= playerMaxLevelNonCHMount it will act like a normal pet.   If the pet mount is above playerMaxLevelNonCHMount, it will not respond to attack type commands and will not be attack by PvE Ai.