10-13-24 [SWGDL] Added armorPSGDecayRateMultiplier / armorDecayRateMultiplier / weapondDecayRateChance / weaponDecayRatewithPowerUpChance / weaponCrystalDamagePerTick / weaponDamagePerTicket / Changed global_variables.lua to be live changes.
For example if you change the sampling rate in global_variables.lua, the system will recognize a change was made and update the live samplingrate values.

I'm sure there are some values that could be changed and it cause a problem.  I will try to document these changes as i run into them.